requirements.txt
Django==5.2.6
django-environ==0.12.0
psycopg2-binary==2.9.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.1
django-autoslug==1.9.9
django-phonenumber-field==8.1.0
phonenumbers==9.0.3
django-countries==7.6.1
pillow==11.3.0
django-filter==25.1
redis==6.0.0
celery==5.5.3
django-celery-beat==2.8.1
flower==2.0.1
django-celery-email==3.0.0
argh==0.31.3
pyYAML==6.0.2
watchdog==6.0.0
resend==2.15.0
channels==4.3.1
daphne==4.2.1
channels_redis==4.3.0
gunicorn==23.0.0
pandas==2.3.3
numpy==2.3.5
scipy==1.16.3
scikit-learn==1.8.0
scikit-surprise==1.1.4